requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dbus-python==1.3.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pillow==11.0.0
PyGObject==3.48.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==68.1.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cu126
torchinfo==1.8.0
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.42.0
